refactor: Move function to dedicated module
Getting log files is part of repair and troubleshooting
